move Tile class into tile module
